bug fixing of test outputs (third time)

+++ src/mapper/Edge.java
@@ -6,14 +6,14 @@ package mapper;
 public class Edge {
     public Integer node1;
     public Integer node2;
-    public Integer weight;
+    public Double weight;
 
     public Edge(Integer node1, Integer node2){
         this.node1 = node1;
         this.node2 = node2;
     }
 
-    public Edge(Integer node1, Integer node2, Integer weight) {
+    public Edge(Integer node1, Integer node2, Double weight) {
         this(node1, node2);
         this.weight = weight;
     }
